Added a new scene and added new tile sets for forest and volcano
I started a new scene for the volcano biome.  So far I just compied over the essentials and got a new terrain going.  As for tiles I got the tile sets for volcano and forest, got them to the right size and all have the needed scripts attached.  Completely good to go.